Move pause gating into worker/pause_gate.rs, use try_fetch_config, and update CHANGELOG.
